@@ -1,6 +1,6 @@
 module antrea.io/nephe
 
-go 1.17
+go 1.19
 
 require (
 	antrea.io/antrea v1.8.0
