@@ -58,7 +58,9 @@ func (c *azureCloud) ProviderType() cloudcommon.ProviderType {
 }
 
 // /////////////////////////////////////////////
-// 	ComputeInterface Implementation
+//
+//	ComputeInterface Implementation
+//
 // /////////////////////////////////////////////.
 // Instances returns VM status for all virtualMachines across all accounts of a cloud provider.
 func (c *azureCloud) Instances() ([]*v1alpha1.VirtualMachine, error) {
@@ -82,7 +84,9 @@ func (c *azureCloud) IsVirtualPrivateCloudPresent(vpcUniqueIdentifier string) bo
 }
 
 // ////////////////////////////////////////////////////////
-// 	AccountMgmtInterface Implementation
+//
+//	AccountMgmtInterface Implementation
+//
 // ////////////////////////////////////////////////////////
 // AddProviderAccount adds and initializes given account of a cloud provider.
 func (c *azureCloud) AddProviderAccount(client client.Client, account *v1alpha1.CloudProviderAccount) error {
